added run simulation button and replaced module
